changed taskId to int64 taskId is set to time.Now().UnixNano() in the client by default added tests for proxying

+++ statement.go
@@ -41,7 +41,7 @@ type Statement struct {
 	functionArgs []Value
 
 	// TaskId determines query task id. (Optional)
-	TaskId int
+	TaskId int64
 
 	// determines if the query should return data
 	returnData bool
